+++ frontend/src/app/[locale]/(base)/vision/_components/sections/VisionHeader.tsx
