+++ frontend/src/app/[locale]/(base)/vision/_components/sections/VisionMission.tsx
@@ -1,7 +1,8 @@
+import VisionPageSection from "src/app/[locale]/(base)/vision/_components/visionPageSection";
+
 import { useMessages, useTranslations } from "next-intl";
 
 import { USWDSIcon } from "src/components/USWDSIcon";
-import VisionPageSection from "src/components/vision/visionPageSection";
 
 export default function VisionMission() {
   const t = useTranslations("Vision.sections.mission");
