+++ frontend/src/app/[locale]/(base)/vision/_components/visionPageSection.tsx
